Backend/requirements.txt
openai>=1.30.0
langchain
langchain-community
langchain-openai
langchain-qdrant
qdrant-client
langchain-text-splitters
beautifulsoup4
aiohttp
lxml
requests==2.32.4
gunicorn
httpx>=0.27.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
fastapi-mail==1.4.1
pydantic[email]==2.7.1
fastapi==0.111.0
python-dotenv
passlib
pathlib
pydantic
pydantic_core
PyJWT
pymongo
requests
uvicorn
email_validator

